+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/internal/UserImpl.kt
@@ -24,6 +24,7 @@ import io.realm.kotlin.internal.platform.freeze
 import io.realm.kotlin.internal.util.use
 import io.realm.kotlin.mongodb.AuthenticationProvider
 import io.realm.kotlin.mongodb.Credentials
+import io.realm.kotlin.mongodb.Functions
 import io.realm.kotlin.mongodb.User
 import io.realm.kotlin.mongodb.UserIdentity
 import io.realm.kotlin.mongodb.auth.ApiKeyAuth
@@ -57,6 +58,8 @@ public class UserImpl(
         get() = RealmInterop.realm_user_get_refresh_token(nativePointer)
     override val deviceId: String
         get() = RealmInterop.realm_user_get_device_id(nativePointer)
+    override val functions: Functions by lazy { FunctionsImpl(app, this) }
+
     override val identities: List<UserIdentity>
         get() = RealmInterop.realm_user_get_all_identities(nativePointer).map {
             UserIdentity(it.id, getProviderFromCore(it.provider))